records: fix inconsistency in api and serializers
* Moves the files logic to a mixin.

* Fixes the inconsistency with the citation table.

* Fixes the ``pre`` and ``post`` dump for the ``_ui_display`` field.
