+++ docker-compose.core.yml
@@ -21,6 +21,7 @@ services:
       service: db
     ports:
      - "5432:5432"
+    command: postgres -c 'max_connections=200'
   mq:
     extends:
       file: docker-compose.services.yml
