+++ inspirehep/records/indexer/base.py
@@ -23,7 +23,7 @@ class InspireRecordIndexer(RecordIndexer):
     """Extend Invenio indexer to properly index Inspire records"""
 
     def _prepare_record(self, record, index, doc_type):
-        data = record.dumps_for_es()
+        data = record.serialize_for_es()
         before_record_index.send(
             current_app._get_current_object(),
             json=data,
