+++ tests/integration/records/views/test_views_experiments.py
@@ -61,7 +61,7 @@ def test_experiments_search_json_get(api_client, db, create_record_factory):
 def test_experiments_record_search_results(api_client, db, es_clear, create_record):
     record = create_record("exp")
 
-    expected_metadata = record._dump_for_es()
+    expected_metadata = record.serialize_for_es()
 
     result = api_client.get("/experiments")
     assert result.json["hits"]["total"] == 1
