+++ tests/integration/records/views/test_views_journals.py
@@ -60,7 +60,7 @@ def test_journals_search_json_get(api_client, db, create_record_factory):
 def test_journal_record_search_results(api_client, db, es_clear, create_record):
     record = create_record("jou")
 
-    expected_metadata = record._dump_for_es()
+    expected_metadata = record.serialize_for_es()
 
     result = api_client.get("/journals")
     assert result.json["hits"]["total"] == 1
